+++ typeClassDefs/wb_walk2014_controlstatus.h
@@ -27,9 +27,9 @@ struct Odometry                         ///< needs to mimic UNSW odometry!
 #ifdef __cplusplus
    /** Comparison operator for Odometry objects */
    inline bool operator== (const Odometry& a) {
-      return (	fabs(forward - a.forward) < DBL_EPSILON &&
-		fabs(left - a.left) < DBL_EPSILON &&
-		fabs(turn - a.turn) < DBL_EPSILON);
+      return (	fabs(static_cast<float>(forward - a.forward)) < DBL_EPSILON &&
+		fabs(static_cast<float>(left - a.left)) < DBL_EPSILON &&
+		fabs(static_cast<float>(turn - a.turn)) < DBL_EPSILON);
    }
 #endif
